Fix errors with #if directives in iOS/macOS projects
